避免轉型浪費時間
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1552 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/clientpackets/C_Attr.java
@@ -586,9 +586,8 @@ public class C_Attr extends ClientBasePacket {
 	}
 
 	// ■■■■■■■■■■■■■ 面向關連 ■■■■■■■■■■■
-	private static final byte HEADING_TABLE_X[] = Base.HEADING_TABLE_X;
-
-	private static final byte HEADING_TABLE_Y[] = Base.HEADING_TABLE_Y;
+	private static final int HEADING_TABLE_X[] = Base.HEADING_TABLE_X;
+	private static final int HEADING_TABLE_Y[] = Base.HEADING_TABLE_Y;
 
 	private void callClan(L1PcInstance pc) {
 		L1PcInstance callClanPc = (L1PcInstance) L1World.getInstance().findObject(pc.getTempID());
