+++ src/net/l1j/server/serverpackets/S_MoveCharPacket.java
@@ -28,8 +28,8 @@ public class S_MoveCharPacket extends ServerBasePacket {
 	private byte[] _byte = null;
 
 	// ■■■■■■■■■■■■■ 移動關連 ■■■■■■■■■■■
-	private static final byte HEADING_TABLE_X[] = Base.HEADING_TABLE_X;
-	private static final byte HEADING_TABLE_Y[] = Base.HEADING_TABLE_Y;
+	private static final int HEADING_TABLE_X[] = Base.HEADING_TABLE_X;
+	private static final int HEADING_TABLE_Y[] = Base.HEADING_TABLE_Y;
 
 	public S_MoveCharPacket(L1Character cha) {
 		int x = cha.getX() - HEADING_TABLE_X[cha.getHeading()];
